refactor: Create packages to organize dto
