+++ api/src/main/java/grooteogi/controller/HashtagController.java
@@ -1,7 +1,7 @@
 package grooteogi.controller;
 
 import grooteogi.domain.Hashtag;
-import grooteogi.dto.HashtagDto;
+import grooteogi.dto.hashtag.HashtagDto;
 import grooteogi.response.BasicResponse;
 import grooteogi.service.HashtagService;
 import java.util.List;
