+++ api/src/main/java/grooteogi/controller/PostController.java
@@ -3,8 +3,8 @@ package grooteogi.controller;
 import grooteogi.domain.Post;
 import grooteogi.dto.PostDto;
 import grooteogi.response.BasicResponse;
-import grooteogi.service.PostService;
 import grooteogi.response.CursorResult;
+import grooteogi.service.PostService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
