+++ api/src/main/java/grooteogi/dto/hashtag/HashtagDto.java
@@ -1,4 +1,4 @@
-package grooteogi.dto;
+package grooteogi.dto.hashtag;
 
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
