+++ api/src/main/java/grooteogi/dto/user/ProfileDto.java
@@ -1,4 +1,4 @@
-package grooteogi.dto;
+package grooteogi.dto.user;
 
 import lombok.Getter;
